CPP layout should not add the root dir of the workspace as a go module
One of the premises of the CPP layout was to avoid relying on the name
of the repository, so that you could have multiple git worktrees or
local clones.

It did this by relying on the go modules declared in the RULES directory
instead.

However, it is still adding the top level directory name as well, which
does not make sense and hinders reproducibility. This change makes sure
only go modules declared in RULES dirs are actually emplaced.